+++ src/components/GoalsComponents/GoalConfigModal/ConfigGoal/BudgetAccordian/BudgetWeekSlider.tsx
@@ -1,58 +1,21 @@
 import { Slider } from "antd";
-import React, { useEffect, useState } from "react";
+import React from "react";
 
 interface IBudgetWeekSlider {
   perWeekHrs: number[];
-  perDayHrs: number[];
-  setPerWeekHrs: (value: number[]) => void;
-  setPerDayHrs: (value: number[]) => void;
-  onDays: string[];
+  budgetPerWeekSummary: string;
+  minWeekValue: number;
+  maxWeekValue: number;
+  handleWeekSliderChange: (val: number[]) => void;
 }
 
 const BudgetWeekSlider: React.FC<IBudgetWeekSlider> = ({
   perWeekHrs,
-  perDayHrs,
-  setPerWeekHrs,
-  setPerDayHrs,
-  onDays,
+  budgetPerWeekSummary,
+  minWeekValue,
+  maxWeekValue,
+  handleWeekSliderChange,
 }) => {
-  const [budgetPerWeekSummary, setBudgetPerWeekSummary] = useState<string>("");
-  const numberOfDays = onDays.length;
-
-  const minWeekValue = perDayHrs[0] * numberOfDays;
-  const maxWeekValue = perDayHrs[1] * numberOfDays;
-
-  const handleWeekSliderChange = (value: number[]) => {
-    let adjustedValue: number[] = value.slice();
-
-    adjustedValue[0] = Math.max(adjustedValue[0], minWeekValue);
-    adjustedValue[1] = Math.min(adjustedValue[1], maxWeekValue);
-
-    if (adjustedValue[0] > adjustedValue[1]) {
-      [adjustedValue[0], adjustedValue[1]] = [adjustedValue[1], adjustedValue[0]];
-    }
-
-    if (perDayHrs[0] === perDayHrs[1]) {
-      adjustedValue = [minWeekValue, maxWeekValue];
-    }
-
-    adjustedValue = adjustedValue.map((val) => Math.max(minWeekValue, Math.min(val, maxWeekValue)));
-    setPerWeekHrs(adjustedValue);
-  };
-
-  useEffect(() => {
-    handleWeekSliderChange(perWeekHrs);
-  }, [perDayHrs, setPerDayHrs, onDays]);
-
-  useEffect(() => {
-    const summary =
-      perWeekHrs[0] === perWeekHrs[1]
-        ? `${perWeekHrs[0]} hrs / week`
-        : `${perWeekHrs[0]} - ${perWeekHrs[1]} hrs / week`;
-
-    setBudgetPerWeekSummary(summary);
-  }, [perWeekHrs]);
-
   return (
     <div>
       <span>{budgetPerWeekSummary}</span>
